library-overdue-reminder/requirements.txt
Flask>=3.0.3,<3.1
Werkzeug>=3.0.3
itsdangerous>=2.2
click>=8.1
Flask-Cors>=4.0.1,<5
SQLAlchemy>=2.0.35,<2.1
python-dotenv>=1.0.1,<2
email-validator>=2.2.0,<3
APScheduler>=3.10.4,<3.11
pytz>=2024.1
tzlocal>=5.2

